requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-WTF
Werkzeug
flask-migrate
geopy
authlib
Flask-Dance
google-auth google-auth-oauthlib requests
mollie-api-python